requirements.txt
anthropic==0.49.0
python-dotenv==1.1.0
streamlit==1.44.1
# Google Generative AI and LangChain dependencies
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.16 
Werkzeug==3.1.3
Flask==3.1.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain_google_genai>=0.0.5
pydantic>=2.0.0
# AutoGen
pyautogen